requirements.txt
google-cloud-bigquery==3.11.0
instaloader==4.10.3
google-api-python-client==2.95.0
pandas==2.0.3
python-dotenv==1.0.0
streamlit==1.28.0
requests==2.31.0
beautifulsoup4==4.12.2
plotly==5.17.0
db-dtypes==1.2.0
